requirements.txt
torch
torchvision
numpy
pandas
matplotlib
scikit-learn
opencv-python
streamlit
tqdm
seaborn
Pillow